% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova.mars}
\alias{anova.mars}
\title{Multivariate Adaptive Regression Splines (MARS)}
\usage{
\method{anova}{mars}(object)
}
\arguments{
\item{object}{an R formula}
}
\description{
Multivariate Adaptive Regression Splines (MARS)
}
