% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Multivariate Adaptive Regression Splines (MARS)}
\usage{
predict(object, newdata)
}
\arguments{
\item{object}{an R formula}

\item{newdata}{a data frame containing the data for the model}
}
\value{
an object of class 'mars'
}
\description{
Multivariate Adaptive Regression Splines (MARS)
}
